feat: smart-download
Only download if the torrent not in the queue or download list.

If the anime type is undetected, then add this to the queue.
If the anime preference is exists, check is the torrent release group priority, download only if it has higher priority
If the anime preference is not exist, then create one